Half edges with opposite sign lasso coefficients not zeroed out.
Instead, allow them to cancel against each other when the half edges are summed into a single edge.
This will better handle the case where one half edge has a large magnitude coefficient and its mate is very close to zero but of the opposite sign.
More pathological cases of half edges with opposite signed coefficients of approximately the same magnitude will be diminished and very unlikely to remain above the threshold to be included.